Time delay between tests added in order to prevent HTTP error 429
