+++ src/test/java/com/example/UserCreationNegativeTest.java
@@ -1,10 +1,11 @@
-package negative;
+package com.example;
 
 import com.example.clients.UserClient;
 import com.example.models.User;
 import com.example.providers.CredentialsProvider;
 import com.example.providers.UserProvider;
 import io.restassured.response.ValidatableResponse;
+import org.junit.After;
 import org.junit.Before;
 import org.junit.Test;
 import org.junit.runner.RunWith;
@@ -24,7 +25,10 @@ public class UserCreationNegativeTest {
     public void setUp() {
         userClient = new UserClient();
     }
-
+    @After
+    public void tearDown() throws InterruptedException {
+        Thread.sleep(300); //In order to avoid 429 Error (too many requests)
+    }
     public UserCreationNegativeTest(User user, int statusCode, String message) {
         this.user = user;
         this.statusCode = statusCode;
