+++ src/test/java/com/example/UserCreationTest.java
@@ -1,4 +1,4 @@
-package usercreation;
+package com.example;
 
 import com.example.clients.UserClient;
 import com.example.models.Credentials;
@@ -6,6 +6,7 @@ import com.example.models.User;
 import com.example.providers.CredentialsProvider;
 import com.example.providers.UserProvider;
 import io.restassured.response.ValidatableResponse;
+import org.junit.After;
 import org.junit.Assert;
 import org.junit.Before;
 import org.junit.Test;
@@ -24,7 +25,10 @@ public class UserCreationTest {
         user = UserProvider.getDefault();
         credentials = CredentialsProvider.getDefault();
     }
-
+    @After
+    public void tearDown() throws InterruptedException {
+        Thread.sleep(300);
+    }
     @Test
     public void userCanBeCreated(){
         ValidatableResponse responseCreate = userClient.create(user);
