+++ src/test/java/com/example/UserLoginTest.java
@@ -1,4 +1,4 @@
-package userlogin;
+package com.example;
 
 import com.example.clients.UserClient;
 import com.example.models.Credentials;
@@ -6,6 +6,7 @@ import com.example.models.User;
 import com.example.providers.CredentialsProvider;
 import com.example.providers.UserProvider;
 import io.restassured.response.ValidatableResponse;
+import org.junit.After;
 import org.junit.Assert;
 import org.junit.Before;
 import org.junit.Test;
@@ -23,6 +24,10 @@ public class UserLoginTest {
         userClient = new UserClient();
         user = UserProvider.getDefault();
         credentials = CredentialsProvider.getDefault();
+    }
+    @After
+    public void tearDown() throws InterruptedException {
+        Thread.sleep(300);
     }
 
     @Test
